ToolsUI de-clutter. Split off more OpPanels, part 4.
