+++ ui/src/main/java/ucar/nc2/ui/op/BufrTableBViewer.java
@@ -29,12 +29,15 @@
  * NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION
  * WITH THE ACCESS, USE OR PERFORMANCE OF THIS SOFTWARE.
  */
-package ucar.nc2.ui;
+package ucar.nc2.ui.op;
 
 import ucar.nc2.constants.CDM;
 import ucar.nc2.ui.dialog.BufrBCompare;
-import ucar.nc2.ui.widget.*;
+import ucar.nc2.ui.widget.BAMutil;
+import ucar.nc2.ui.widget.FileManager;
+import ucar.nc2.ui.widget.IndependentWindow;
 import ucar.nc2.ui.widget.PopupMenu;
+import ucar.nc2.ui.widget.TextHistoryPane;
 import ucar.nc2.util.Misc;
 import ucar.unidata.util.StringUtil2;
 import ucar.util.prefs.PreferencesExt;
@@ -48,10 +51,6 @@ import ucar.unidata.io.RandomAccessFile;
 import javax.swing.*;
 import javax.swing.event.ListSelectionListener;
 import javax.swing.event.ListSelectionEvent;
-
-import ucar.nc2.ui.widget.TextHistoryPane;
-import ucar.nc2.ui.widget.FileManager;
-import ucar.nc2.ui.widget.BAMutil;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
