+++ ui/src/main/java/ucar/nc2/ui/op/CollectionSpecTable.java
@@ -16,8 +16,6 @@ import ucar.util.prefs.ui.BeanTable;
 
 import javax.swing.*;
 import java.awt.*;
-import java.io.ByteArrayOutputStream;
-import java.io.PrintStream;
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.util.ArrayList;
