+++ ui/src/main/java/ucar/nc2/ui/op/DatasetViewer.java
@@ -7,7 +7,6 @@ package ucar.nc2.ui.op;
 
 import java.awt.*;
 import java.awt.event.ActionEvent;
-import java.awt.event.ActionListener;
 import java.awt.event.WindowAdapter;
 import java.awt.event.WindowEvent;
 import java.beans.PropertyChangeEvent;
