+++ ui/src/main/java/ucar/nc2/ui/op/DatasetWriterPanel.java
@@ -8,16 +8,13 @@ package ucar.nc2.ui.op;
 import ucar.nc2.NetcdfFile;
 import ucar.nc2.ui.OpPanel;
 import ucar.nc2.ui.ToolsUI;
-import ucar.nc2.ui.widget.BAMutil;
 import ucar.util.prefs.PreferencesExt;
 
 import java.awt.BorderLayout;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.io.StringWriter;
 import java.lang.invoke.MethodHandles;
-import java.util.Formatter;
 import javax.swing.JSplitPane;
 
 /**
