+++ ui/src/main/java/ucar/nc2/ui/op/package.html
@@ -1,0 +1,9 @@
+<!doctype html public "-//w3c//dtd html 4.0 transitional//en">
+<html>
+<head>
+   <title>package ucar.nc2.ui.op</title>
+</head>
+<body bgcolor="#FFFFFF">
+<p>NetCDF Tools User Interface Op Panels</p>
+</body>
+</html>
